implement vertical nav for detail view
