+++ services/ui-src/src/components/PortalTable.tsx
@@ -1,6 +1,15 @@
 import React, { ReactNode } from "react";
 import {
+  HeaderGroup,
+  TableInstance,
   TableOptions,
+  UseExpandedInstanceProps,
+  UseExpandedRowProps,
+  UseFiltersInstanceProps,
+  UseGlobalFiltersInstanceProps,
+  UseSortByColumnProps,
+  UseSortByInstanceProps,
+  Row,
   useExpanded,
   useFilters,
   useGlobalFilter,
@@ -53,20 +62,15 @@ export default function PortalTable<V extends {} = {}>({
     getTableProps,
     getTableBodyProps,
     headerGroups,
-    // @ts-ignore FIXME remove when react-table types are improved
     preFilteredRows,
-    // @ts-ignore FIXME remove when react-table types are improved
     preGlobalFilteredRows,
     rows,
-    // @ts-ignore FIXME remove when react-table types are improved
     setAllFilters,
-    // @ts-ignore FIXME remove when react-table types are improved
     setGlobalFilter,
     prepareRow,
   } = useTable<V>(
     {
       // don't revert to original sort state when `data` prop changes
-      // @ts-ignore FIXME remove when react-table types are improved
       autoResetSortBy: false,
       // once a column has been sorted, only toggle between sort orders - do
       // not go to original state
@@ -83,7 +87,11 @@ export default function PortalTable<V extends {} = {}>({
     useFilters,
     useSortBy,
     useExpanded
-  );
+  ) as TableInstance<V> &
+    UseFiltersInstanceProps<V> &
+    UseGlobalFiltersInstanceProps<V> &
+    UseSortByInstanceProps<V> &
+    UseExpandedInstanceProps<V>;
 
   return (
     <>
@@ -105,22 +113,21 @@ export default function PortalTable<V extends {} = {}>({
             {headerGroups.map((headerGroup) => (
               <tr {...headerGroup.getHeaderGroupProps()}>
                 {expandable && <th />}
-                {headerGroup.headers.map((column) => (
+                {(
+                  headerGroup.headers as (HeaderGroup<V> &
+                    UseSortByColumnProps<V>)[]
+                ).map((column) => (
                   <th
-                    // @ts-ignore FIXME remove when react-table types are improved
                     {...column.getHeaderProps(column.getSortByToggleProps())}
                     id={`${column.id}ColHeader`}
                   >
                     {column.render("Header")}
                     <span className="sort-icons-table">
-                      {/* @ts-ignore FIXME remove when react-table types are improved */}
                       {column.isSorted ? (
                         <FontAwesomeIcon
-                          // @ts-ignore FIXME remove when react-table types are improved
                           icon={column.isSortedDesc ? faSortDown : faSortUp}
                         />
                       ) : (
-                        // @ts-ignore FIXME remove when react-table types are improved
                         column.canSort && <FontAwesomeIcon icon={faSort} />
                       )}
                     </span>
@@ -131,50 +138,47 @@ export default function PortalTable<V extends {} = {}>({
           </thead>
           {rows.length > 0 && (
             <tbody {...getTableBodyProps()}>
-              {rows.map((row, rowIndex) => {
-                // @ts-ignore FIXME remove when react-table types are improved
-                prepareRow(row, rowIndex);
-                return (
-                  <tr
-                    {...row.getRowProps()}
-                    className={cx({
-                      // @ts-ignore FIXME remove when react-table types are improved
-                      "parent-row-expanded": row.isExpanded,
-                      // @ts-ignore FIXME remove when react-table types are improved
-                      "child-row-expanded": row.depth > 0,
-                    })}
-                  >
-                    {expandable && (
-                      <td className="row-expander-cell">
-                        {/* @ts-ignore */}
-                        {row.depth === 0 && (
-                          <button
-                            aria-label="Expand row"
-                            // @ts-ignore FIXME remove when react-table types are improved
-                            disabled={!row.canExpand}
-                            // @ts-ignore FIXME remove when react-table types are improved
-                            onClick={() => row.toggleRowExpanded()}
-                          >
-                            <FontAwesomeIcon icon={faChevronDown} />
-                          </button>
-                        )}
-                      </td>
-                    )}
-                    {row.cells.map((cell, index) => {
-                      return (
-                        <td
-                          id={
-                            headerGroups[0].headers[index].id + "-" + rowIndex
-                          }
-                          {...cell.getCellProps()}
-                        >
-                          {cell.render("Cell")}
+              {(rows as (Row<V> & UseExpandedRowProps<V>)[]).map(
+                (row, rowIndex) => {
+                  // @ts-ignore FIXME remove when react-table types are improved
+                  prepareRow(row, rowIndex);
+                  return (
+                    <tr
+                      {...row.getRowProps()}
+                      className={cx({
+                        "parent-row-expanded": row.isExpanded,
+                        "child-row-expanded": row.depth > 0,
+                      })}
+                    >
+                      {expandable && (
+                        <td className="row-expander-cell">
+                          {row.depth === 0 && (
+                            <button
+                              aria-label="Expand row"
+                              disabled={!row.canExpand}
+                              onClick={() => row.toggleRowExpanded()}
+                            >
+                              <FontAwesomeIcon icon={faChevronDown} />
+                            </button>
+                          )}
                         </td>
-                      );
-                    })}
-                  </tr>
-                );
-              })}
+                      )}
+                      {row.cells.map((cell, index) => {
+                        return (
+                          <td
+                            id={
+                              headerGroups[0].headers[index].id + "-" + rowIndex
+                            }
+                            {...cell.getCellProps()}
+                          >
+                            {cell.render("Cell")}
+                          </td>
+                        );
+                      })}
+                    </tr>
+                  );
+                }
+              )}
             </tbody>
           )}
         </table>
